Added full support for standard and extended can message id
- Fixed a bug where data was written using an invalid NULL pointer

+++ src/shell_can.h
@@ -11,7 +11,7 @@ extern "C" {
 struct can_command
 {
         char *cmd;
-        unsigned int opt;
+        long unsigned int opt;
         union
         {
                 struct {
